Major main render loop & shading refactoring
Raytracer :
    - [UPDATE] computeIllumination now sample incoming light like so :
        - Indirect light diffusion -> through Shader.sampleDiffusion()
        - Direct lighting -> loop over lights & compute incoming lighting (using absorbance from shader.sampleDiffusion())
        - Reflection -> through computeReflection()
        - Transmission (refracted light) -> through computeReflection()

Shader :
    - [UPDATE] split scatter() method in 3 separate sampleDiffusion(), sampleReflection() & sampleTransmission()
               Each one of this method returns a single absorbance & outRay that will be used to compute incoming lighting at hit position

Light :
    - [ADD] getInfiniteIllumination() used to get illumination when a ray didn't hit anything
    - [ADD] DirectionalLight & EnvironmentLight classes

Ray :
    [RENAME] shader_ptr -> shader

+++ bounceRenderer/src/FileAdapters.cpp
@@ -3,11 +3,11 @@
 
 bool PPMAdapter::write(RayTracer& tracer) const
 {
-    // opening file stream
+    // open file stream
     std::ofstream outFileStream(tracer.outpath());
 
-    float width = tracer.camera().width();
-    float height = tracer.camera().height();
+    unsigned int width = tracer.camera().width();
+    unsigned int height = tracer.camera().height();
     // Write ppm format data
     outFileStream << "P3\n" << width << " " << height << "\n255\n";
